+++ src/modules/map/store/gettersMap.js
@@ -1,10 +1,19 @@
 import stateMap from "./stateMap";
+import {MapMode} from "./enums";
 import {generateSimpleGetters} from "../../../app-store/utils/generators";
 import {createGfiFeature} from "../../../api/gfi/getWmsFeaturesByMimeType";
 import {getGfiFeaturesByTileFeature} from "../../../api/gfi/getGfiFeaturesByTileFeature";
 
 const gettersMap = {
     ...generateSimpleGetters(stateMap),
+
+    /**
+     * @param {Object} state the map state
+     * @return {boolean} whether the portal is currently in 3D mode
+     */
+    is3d ({mapMode}) {
+        return mapMode === MapMode.MODE_3D;
+    },
 
     /**
      * gets all visible layers
@@ -133,7 +142,7 @@ const gettersMap = {
      * @param {Object} g getters
      * @returns {Function} layer getter by id
      */
-    layerById: (_, g) => id => g.layers[id],
+    layerById: (_, g) => id => g.layers && g.layers[id],
     /**
      * @param {Object} _ state
      * @param {Object} g getters