+++ src/modules/tools/measure/util/measureSource.js
@@ -1,0 +1,3 @@
+import VectorSource from "ol/source/Vector.js";
+
+export default new VectorSource();